refactor: centralize UDF compilation logic, improve capture analysis, and enhance function resolution
