+++ pkg/compiler/visitor.go
@@ -30,11 +30,7 @@ func (v *Visitor) VisitProgram(ctx *fql.ProgramContext) interface{} {
 		v.VisitHead(head.(*fql.HeadContext))
 	}
 
-	v.Ctx.UDFs = internal.CollectUDFs(v.Ctx, ctx)
-	if v.Ctx.UDFs != nil {
-		v.Ctx.UDFScope = v.Ctx.UDFs.GlobalScope
-	}
-
+	v.Ctx.UDFCompiler.CollectProgram(ctx)
 	v.Ctx.StmtCompiler.Compile(ctx.Body())
 	v.Ctx.UDFCompiler.CompileAll()
